backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1;python_version<"3.9"
Django==2.2
django-cors-headers==3.6.0
djangorestframework==3.12.0
pystache==0.6.0
pytz==2022.4
sqlparse==0.4.3
static==1.1.1
